Edit file for components
