+++ src/components/Book.js
@@ -1,29 +1,20 @@
+/* eslint-disable react/prop-types */
 import React from 'react';
-import PropTypes from 'prop-types';
 
 const Book = (props) => {
-  const {
-    id, title, author, category, handleDeleteProps,
-  } = props;
-  const elementId = `book${id}`;
+  const { book } = props;
   return (
-    <li className="single-book" id={elementId}>
+    <li className="single-book">
       <div className="left-side">
         <div className="book-details">
-          <div className="book-category">{category}</div>
-          <div className="book-title">{title}</div>
-          <div className="book-category">{author}</div>
+          <div className="book-category">{book.category}</div>
+          <div className="book-title">{book.title}</div>
+          <div className="book-category">{book.author}</div>
         </div>
         <div className="book-buttons">
           <button type="button" className="comments-btn">Comments</button>
           <div className="separator">|</div>
-          <button
-            type="button"
-            className="remove-btn"
-            onClick={() => handleDeleteProps(id)}
-          >
-            Remove
-          </button>
+          <button type="button" className="remove-btn">Remove</button>
           <div className="separator">|</div>
           <button type="button" className="edit-btn">Edit</button>
         </div>
@@ -38,19 +29,10 @@ const Book = (props) => {
       <div className="sides-separator" />
       <div className="right-side">
         <div className="current-text">Current chapter</div>
-        <div className="current-value">{17}</div>
+        <div className="current-value">{book.currentChapter}</div>
         <button type="button" className="">Update progress</button>
       </div>
     </li>
   );
 };
-
-Book.propTypes = {
-  id: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-  author: PropTypes.string.isRequired,
-  category: PropTypes.string.isRequired,
-  handleDeleteProps: PropTypes.func.isRequired,
-};
-
 export default Book;
