+++ src/components/Books.js
@@ -1,47 +1,21 @@
+/* eslint-disable react/prop-types */
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { v4 as uuidv4 } from 'uuid';
-import { addBook, removeBook } from '../redux/books/books';
 import Book from './Book';
 import BookForm from './BookForm';
 
-const Books = () => {
-  const books = useSelector((state) => state.booksReducer);
-  const dispatch = useDispatch();
-
-  const addBookToStore = (title, author, category) => {
-    const newBook = {
-      id: uuidv4(),
-      title,
-      author,
-      category,
-    };
-    dispatch(addBook(newBook));
-  };
-
-  const deleteBookFromStore = (id) => {
-    dispatch(removeBook(id));
-  };
-
+const Books = (props = { books: [] }) => {
+  let { books } = props;
+  if (!books) books = [];
   return (
     <>
       <ul className="listContainer">
         {books.map((book) => (
-          <Book
-            key={book.id}
-            id={book.id}
-            title={book.title}
-            author={book.author}
-            category={book.category}
-            handleDeleteProps={deleteBookFromStore}
-          />
+          <Book key={book.id} book={book} />
         ))}
       </ul>
-      <BookForm addBookProps={addBookToStore} />
+      <BookForm />
     </>
   );
 };
-
-Books.propTypes = {};
 
 export default Books;
